+++ packages/spright-components/CHANGELOG.md
@@ -1,8 +1,33 @@
 # Change Log - @ni/spright-components
 
-<!-- This log was last generated on Fri, 27 Jun 2025 15:07:42 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 08 Jul 2025 00:20:19 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 5.5.3
+
+Tue, 08 Jul 2025 00:20:19 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v33.12.1
+
+## 5.5.2
+
+Mon, 07 Jul 2025 16:46:23 GMT
+
+### Patches
+
+- Style improvements for chat conversation ([ni/nimble@69a41ca](https://github.com/ni/nimble/commit/69a41ca15b6f41e3af2b7d5bc2472bc969925777))
+- Bump @ni/nimble-components to v33.12.0
+
+## 5.5.1
+
+Wed, 02 Jul 2025 14:09:36 GMT
+
+### Patches
+
+- Delegate focus to input when chat is clicked ([ni/nimble@44fd8dd](https://github.com/ni/nimble/commit/44fd8dde63ffa41369de86785399c4c1a5eef22f))
 
 ## 5.5.0